Move graph objects to new file
