+++ src/backend/gedcomx-extensions.ts
@@ -2,78 +2,14 @@ import * as GedcomX from "gedcomx-js";
 import {translationToString} from "../main";
 import config from "../config";
 import {
-  DisplayProperties,
   Fact, FamilyView,
   Person,
   PlaceReference,
   Qualifier, Relationship
 } from "gedcomx-js";
+import {GraphFamily, GraphPerson} from "./graph";
 
 GedcomX.enableRsExtensions();
-
-type PersonType = "person";
-type FamilyType = "family" | "etc";
-type GraphObjectType = PersonType | FamilyType;
-
-export interface GraphObject {
-  type: GraphObjectType | `${GraphObjectType}-removed`
-  width: number
-  height: number
-  bounds
-  viewId: number
-}
-
-export class GraphPerson extends GedcomX.DisplayProperties implements GraphObject {
-  type: PersonType | `${PersonType}-removed` = "person"
-  data: Person
-  width = config.gridSize * 5
-  height = config.gridSize / 2 * 2.25
-  bounds
-  viewId
-
-  constructor(person: Person) {
-    super({
-      name: person.getFullName(),
-      gender: readableGender(person.getGender()),
-      ascendancyNumber: getGeneration(person)
-    });
-    this.data = person;
-  }
-
-  equals = (person: Person | GraphPerson | DisplayProperties) => {
-    if (person instanceof DisplayProperties) {
-      return person.getName() === this.getName()
-        && person.getBirthDate() === this.getBirthDate();
-    }
-
-    if (person instanceof GraphPerson) {
-      person = person.data;
-    }
-    return person.getId() === this.data.getId();
-  }
-
-  toString = () => {
-    return `${this.data.getFullName()} (#${this.data.getId()} @${this.viewId})`
-  }
-}
-
-export class GraphFamily extends GedcomX.FamilyView implements GraphObject {
-  type: FamilyType | `${FamilyType}-removed` = "family"
-  width = config.margin * 2
-  height = config.margin * 2
-  bounds
-  viewId
-  marriage
-
-  equals = (object: FamilyView): Boolean => {
-    return this.getParent1().getResource() === object.getParent1().getResource() &&
-      this.getParent2().getResource() === object.getParent2().getResource();
-  }
-
-  toString = () => {
-    return `Family of ${this.getParent1().getResource()} and ${this.getParent2().getResource()}`
-  }
-}
 
 export enum Confidence {
   Low = "http://gedcomx.org/Low", Medium = "http://gedcomx.org/Medium", High = "http://gedcomx.org/High"
@@ -141,17 +77,12 @@ export const baseUri = "http://gedcomx.org/";
 
 // Person
 
-function getGeneration(person): number | undefined {
+export function getGeneration(person): number | undefined {
   let generationFacts = person.getFactsByType(PersonFactTypes.Generation);
   if (!generationFacts.length) {
     return undefined
   }
   return generationFacts[0].value
-}
-
-function readableGender(gender): string {
-  gender = gender || {type: GenderTypes.Unknown};
-  return gender.type.substring(baseUri.length).toLowerCase();
 }
 
 let referenceAge = {
